refactor, add AbstractRetailOutlet class
